MBS-10577 Can apply build checks to Android app without root project
