+++ subprojects/gradle/build-checks/src/main/kotlin/com/avito/android/build_checks/internal/unique_r/UniqueRClassesTaskCreator.kt
@@ -12,7 +12,7 @@ import org.gradle.api.Task
 import org.gradle.api.tasks.TaskProvider
 import org.gradle.kotlin.dsl.register
 
-internal class UniqueRClassesTaskProvider(
+internal class UniqueRClassesTaskCreator(
     private val appProject: Project,
     private val config: AndroidAppCheck.UniqueRClasses
 ) {
